refactor: factor out exportSpace listener
Factor out the exportSpace listener to its own file.

refs #104
