+++ public/app/listeners/index.js
@@ -5,6 +5,7 @@ const getSpaces = require('./getSpaces');
 const deleteSpace = require('./deleteSpace');
 const showSyncSpacePrompt = require('./showSyncSpacePrompt');
 const syncSpace = require('./syncSpace');
+const exportSpace = require('./exportSpace');
 
 module.exports = {
   loadSpace,
@@ -14,4 +15,5 @@ module.exports = {
   showSyncSpacePrompt,
   syncSpace,
   deleteSpace,
+  exportSpace,
 };
